moved random note generation to separate module
